RegionCode update

+++ SelfSign/Controllers/UserController.cs
@@ -256,7 +256,7 @@ namespace SelfSign.Controllers
         
         public string SubDivisionAddress { get; set; }
         public string Gender { get; set; }
-        public int RegionCode { get; set; }
+        public long RegionCode { get; set; }
         public string? Citizenship { get; set; }
         public string? Phone { get; set; }
         public string? Email { get; set; }
